make node happy
