+++ src/classes/builders/bases/index.ts
@@ -2,4 +2,3 @@ export * from './BaseBuilder.js';
 export * from './BaseCommand.js';
 export * from './BaseOption.js';
 export * from './CommandWithOptions.js';
-export * from './OptionWithAutocomplete.js';
