+++ src/classes/builders/commands/index.ts
@@ -1,5 +1,3 @@
-export * from '../bases/BaseCommand.js';
-export * from '../bases/CommandWithOptions.js';
 export * from './ContextMenuCommandBuilder.js';
 export * from './SlashCommandBuilder.js';
 export * from './SlashCommandSubcommandBuilder.js';
