--- src/classes/builders/index.ts
@@ -1,3 +1,0 @@
-export * from './bases';
-export * from './options';
-export * from './commands';
