+++ src/classes/builders/options/AttachmentOption.ts
@@ -1,5 +1,5 @@
 import { SlashCommandAttachmentOption, ApplicationCommandOptionType } from 'discord.js';
-import { BaseOption } from '#builders';
+import { BaseOption } from '#builderBases';
 
 // https://discord.js.org/#/docs/builders/main/class/SlashCommandAttachmentOption
 export class AttachmentOption extends BaseOption<SlashCommandAttachmentOption> {
