+++ src/classes/builders/options/BooleanOption.ts
@@ -1,5 +1,5 @@
 import { SlashCommandBooleanOption, ApplicationCommandOptionType } from 'discord.js';
-import { BaseOption } from '#builders';
+import { BaseOption } from '#builderBases';
 
 // https://discord.js.org/#/docs/builders/main/class/SlashCommandBooleanOption
 export class BooleanOption extends BaseOption<SlashCommandBooleanOption> {
