+++ src/classes/builders/options/IntegerOption.ts
@@ -1,10 +1,13 @@
 import { ApplicationCommandOptionType, SlashCommandIntegerOption } from 'discord.js';
-import { OptionWithAutocomplete } from '#builders';
+import { BaseOption } from '#builderBases';
 
-export class IntegerOption extends OptionWithAutocomplete<SlashCommandIntegerOption> {
+export class IntegerOption extends BaseOption<SlashCommandIntegerOption> {
+    choiceQueue: number[];
+
     constructor(baseKey: string) {
         super(ApplicationCommandOptionType.Integer, baseKey);
         this.builder = new SlashCommandIntegerOption();
+        this.choiceQueue = [];
     }
 
     setMaxValue(num: number) {
@@ -14,6 +17,35 @@ export class IntegerOption extends OptionWithAutocomplete<SlashCommandIntegerOpt
 
     setMinValue(num: number) {
         this.builder.setMinValue(num);
+        return this;
+    }
+
+    setAutocomplete(bool: boolean) {
+        this.builder.setAutocomplete(bool);
+        return this;
+    }
+
+    addChoices(...choices: number[]) {
+        this.choiceQueue = choices;
+        return this;
+    }
+
+    prepareChoices() {
+        const preparedChoices = this.choiceQueue.map((value: number) => {
+            const name = this.getDefault(`${this.baseKey}.CHOICES.${value.toString().toUpperCase()}`);
+            const name_localizations = this.getLocalizations(
+                `${this.baseKey}.CHOICES.${value.toString().toUpperCase()}`
+            );
+
+            return {
+                name,
+                name_localizations,
+                value
+            };
+        });
+
+        this.choiceQueue = [];
+        this.builder.addChoices(...preparedChoices);
         return this;
     }
 }
