+++ src/classes/builders/options/NumberOption.ts
@@ -1,10 +1,13 @@
 import { ApplicationCommandOptionType, SlashCommandNumberOption } from 'discord.js';
-import { OptionWithAutocomplete } from '#builders';
+import { BaseOption } from '#builderBases';
 
-export class NumberOption extends OptionWithAutocomplete<SlashCommandNumberOption> {
+export class NumberOption extends BaseOption<SlashCommandNumberOption> {
+    choiceQueue: number[];
+
     constructor(baseKey: string) {
         super(ApplicationCommandOptionType.Number, baseKey);
         this.builder = new SlashCommandNumberOption();
+        this.choiceQueue = [];
     }
 
     setMaxValue(num: number) {
@@ -14,6 +17,35 @@ export class NumberOption extends OptionWithAutocomplete<SlashCommandNumberOptio
 
     setMinValue(num: number) {
         this.builder.setMinValue(num);
+        return this;
+    }
+
+    setAutocomplete(bool: boolean) {
+        this.builder.setAutocomplete(bool);
+        return this;
+    }
+
+    addChoices(...choices: number[]) {
+        this.choiceQueue = choices;
+        return this;
+    }
+
+    prepareChoices() {
+        const preparedChoices = this.choiceQueue.map((value: number) => {
+            const name = this.getDefault(`${this.baseKey}.CHOICES.${value.toString().toUpperCase()}`);
+            const name_localizations = this.getLocalizations(
+                `${this.baseKey}.CHOICES.${value.toString().toUpperCase()}`
+            );
+
+            return {
+                name,
+                name_localizations,
+                value
+            };
+        });
+
+        this.choiceQueue = [];
+        this.builder.addChoices(...preparedChoices);
         return this;
     }
 }
