+++ src/classes/builders/options/StringOption.ts
@@ -1,10 +1,13 @@
 import { ApplicationCommandOptionType, SlashCommandStringOption } from 'discord.js';
-import { OptionWithAutocomplete } from '#builders';
+import { BaseOption } from '#builderBases';
 
-export class StringOption extends OptionWithAutocomplete<SlashCommandStringOption> {
+export class StringOption extends BaseOption<SlashCommandStringOption> {
+    choiceQueue: string[];
+
     constructor(baseKey: string) {
         super(ApplicationCommandOptionType.String, baseKey);
         this.builder = new SlashCommandStringOption();
+        this.choiceQueue = [];
     }
 
     setMaxLength(num: number) {
@@ -14,6 +17,34 @@ export class StringOption extends OptionWithAutocomplete<SlashCommandStringOptio
 
     setMinLength(num: number) {
         this.builder.setMinLength(num);
+        return this;
+    }
+    setAutocomplete(bool: boolean) {
+        this.builder.setAutocomplete(bool);
+        return this;
+    }
+
+    addChoices(...choices: string[]) {
+        this.choiceQueue = choices;
+        return this;
+    }
+
+    prepareChoices() {
+        const preparedChoices = this.choiceQueue.map((value: string) => {
+            const name = this.getDefault(`${this.baseKey}.CHOICES.${value.toString().toUpperCase()}`);
+            const name_localizations = this.getLocalizations(
+                `${this.baseKey}.CHOICES.${value.toString().toUpperCase()}`
+            );
+
+            return {
+                name,
+                name_localizations,
+                value
+            };
+        });
+
+        this.choiceQueue = [];
+        this.builder.addChoices(...preparedChoices);
         return this;
     }
 }
