+++ src/classes/builders/options/UserOption.ts
@@ -1,4 +1,4 @@
-import { BaseOption } from '#builders';
+import { BaseOption } from '#builderBases';
 import { ApplicationCommandOptionType, SlashCommandUserOption } from 'discord.js';
 
 // https://discord.js.org/#/docs/builders/main/class/SlashCommandUserOption
