+++ tsconfig.json
@@ -8,8 +8,10 @@
         "baseUrl": "src",
         "paths": {
             "#api": ["api/index.ts"],
-            "#builders": ["classes/builders/index.ts"],
+            "#builders": ["classes/builders/commands/index.ts"],
+            "#builderBases": ["classes/builders/bases/index.ts"],
             "#classes": ["classes/index.ts"],
+            "#commands": ["classes/builders/commands/index.ts"],
             "#env": ["env/index.ts"],
             "#handlers": ["classes/handlers/index.ts"],
             "#modules": ["classes/modules/index.ts"],
